add RIOT Metadata

+++ src/main/java/com/krazytop/nomenclature/lol/LOLItemNomenclature.java
@@ -22,7 +22,7 @@ public class LOLItemNomenclature extends LOLNomenclature {
     private int baseGold;
     private int totalGold;
     private List<String> tags;
-    private Map<String, Integer> stats;
+    private Map<String, Double> stats;
     @JsonAlias("into")
     private List<String> toItems;
     @JsonAlias("from")
